added encrypt method to cipher class, read-in module to cipher class and created passing tests for all methods
