+++ test/enigma_test.rb
@@ -64,14 +64,6 @@ class EnigmaTest <Minitest::Test
     assert_equal 53, enigma.d_shift("04853", enigma.get_current_date)
   end
 
-  def test_encrypt
-    enigma = Enigma.new
-
-    expected = {:encryption=>"ozvqmrxzkmxszcpk", :key=>"04853", :date=>"092120"}
-
-    assert_equal expected, enigma.encrypt("garrett cottrell", "04853")
-  end
-
   def test_decrypt
     enigma = Enigma.new
     expected = {:decryption=>"garrett cottrell", :key=>"04853", :date=>"092120"}
